+++ crates/wdl-lsp/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.10.0 - 07-09-2025
+
 #### Added
 
 * Added `goto_definition`, `find_all_references` tests ([#489](https://github.com/stjude-rust-labs/wdl/pull/489)).